add PP3.1.3v2

+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserServiceImpl.java
@@ -5,20 +5,19 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.kata.spring.boot_security.demo.models.Role;
 import ru.kata.spring.boot_security.demo.models.User;
 import ru.kata.spring.boot_security.demo.repositories.UserRepository;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public class UserServiceImpl implements UserService{
+public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
     private final PasswordEncoder passwordEncoder;
+
     @Autowired
     public UserServiceImpl(UserRepository userRepository, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
@@ -30,10 +29,11 @@ public class UserServiceImpl implements UserService{
         return userRepository.findAll();
     }
 
+
     @Override
     public User getUserById(long id) {
-        Optional<User>user = userRepository.findById(id);
-        if(user.isEmpty()){
+        Optional<User> user = userRepository.findById(id);
+        if (user.isEmpty()) {
             throw new UsernameNotFoundException("Нет пользователя с указанным Id");
         }
         return user.get();
@@ -55,23 +55,15 @@ public class UserServiceImpl implements UserService{
     }
 
     @Override
-    @Transactional
-    public void deleteById(long id) {
-        userRepository.deleteById(id);
+    public void deleteUserById(long id) {
+        if (userRepository.findById(id).isPresent())
+            userRepository.deleteById(id);
     }
 
     @Override
     @Transactional
-    public void changePassword(long id, String password) {
-        userRepository.findById(id).ifPresent(user -> user.setPassword(passwordEncoder.encode(password)));
-    }
-    @Override
-    @Transactional
-    public void register(User user){
+    public void register(User user) {
         user.setPassword(passwordEncoder.encode(user.getPassword()));
-//        List<Role> roles = new ArrayList<>();
-//        roles.add(new Role("ROLE_USER"));
-//        user.setRoles(roles);
         userRepository.save(user);
     }
 }
